+++ source/Sailfish.TestAdapter/Display/TestOutputWindow/Table.cs
@@ -1,0 +1,3 @@
+namespace Sailfish.TestAdapter.Display.TestOutputWindow;
+
+public record Table(string Name, double Before, double After);