+++ source/Sailfish/Contracts.Public/Notifications/TestCaseCompletedNotification.cs
@@ -1,9 +1,12 @@
 using MediatR;
+using Sailfish.Contracts.Public.Models;
 using Sailfish.Contracts.Public.Serialization.Tracking.V1;
+using System.Collections.Generic;
 
 namespace Sailfish.Contracts.Public.Notifications;
 
-public class TestCaseCompletedNotification(ClassExecutionSummaryTrackingFormat testCaseExecutionResult) : INotification
-{
-    public ClassExecutionSummaryTrackingFormat TestCaseExecutionResult { get; } = testCaseExecutionResult;
-}
+public record TestCaseCompletedNotification(
+    ClassExecutionSummaryTrackingFormat ClassExecutionSummaryTrackingFormat,
+    TestInstanceContainerExternal TestInstanceContainerExternal,
+    IEnumerable<dynamic> TestCaseGroup
+) : INotification;