+++ source/Sailfish/Contracts.Public/Notifications/TestCaseExceptionNotification.cs
@@ -1,0 +1,8 @@
+﻿using MediatR;
+using Sailfish.Contracts.Public.Models;
+using System;
+using System.Collections.Generic;
+
+namespace Sailfish.Contracts.Public.Notifications;
+
+public record TestCaseExceptionNotification(TestInstanceContainerExternal? TestInstanceContainer, IEnumerable<dynamic> TestCaseGroup, Exception? Exception) : INotification;